[SM64] Bring more props to combined and merge level enums
There is only one level enum (which I took the time to organize), binary will simply error out with Custom (tho it is not fully out of the possibility to support custom levels in binary in the future), upgrade code added.
Dl and animation now share props with combined, animation rework actually uses these but for now both just redraw the prop to keep the ui the same.
Imports and address converter left untouched besides enum change, these should not be per scene.
Convert to str in getPathAndLevel to avoid issues with posix paths -> abspath

Alternatively, should this be just a pr for level enums?
Closes #498