[SM64] Concept: Use desgraph updates to set current versions at runtime
Instead of always running upgrade code on new objects/bones/scenes on file load, we make sure they get the new version as soon as possible. This is ideal considering we are gathering properties from keys that could be very likely exist in another addon.

That said, I feel like there has to be a better approach to this? I can´t really figure out one. This does work it just feels hacky I guess, even if this is how materials work.